Fix outdated removing plugin order
